feat: add game repository structure

+++ src/main/java/com/pedro/gamecatalogs/repository/GameRepository.java
@@ -3,5 +3,5 @@ package com.pedro.gamecatalogs.repository;
 import com.pedro.gamecatalogs.model.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public class GameRepository  extends JpaRepository<Game, Integer> {
+public class GameRepository  extends JpaRepository<Game, Long> {
 }
